@@ -3,12 +3,12 @@
     "description": "Provides assertions for unit tests.",
     "license": "BSD-3-Clause",
     "require": {
-        "php": "^7.0",
-        "sebastian/comparator": "^1.2",
-        "sebastian/exporter": "^2.0"
+        "php": "^7.1",
+        "sebastian/comparator": "^2.0",
+        "sebastian/exporter": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7",
+        "phpunit/phpunit": "^6.3",
         "mikey179/vfsStream": "^1.6"
     },
     "autoload": {
@@ -22,14 +22,9 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.2.x-dev"
+            "dev-master": "3.0.x-dev"
         }
     },
     "minimum-stability": "dev",
-    "prefer-stable": true,
-    "config": {
-        "platform": {
-            "php": "7.0.0"
-        }
-    }
+    "prefer-stable": true
 }
