@@ -6,11 +6,11 @@
  * file that was distributed with this source code.
  */
 namespace bovigo\assert;
-if (class_exists('PHPUnit_Framework_ExpectationFailedException')) {
+if (class_exists('\PHPUnit\Framework\ExpectationFailedException')) {
     /**
      * Thrown when amount of received arguments is lower than expected amount.
      */
-    class AssertionFailure extends \PHPUnit_Framework_ExpectationFailedException
+    class AssertionFailure extends \PHPUnit\Framework\ExpectationFailedException
     {
         // intentionally empty
     }
@@ -23,4 +23,3 @@ if (class_exists('PHPUnit_Framework_ExpectationFailedException')) {
         // intentionally empty
     }
 }
-
