@@ -8,6 +8,8 @@ declare(strict_types=1);
  */
 namespace bovigo\assert\predicate;
 use bovigo\assert\AssertionFailure;
+use PHPUnit\Framework\TestCase;
+
 use function bovigo\assert\assert;
 use function bovigo\assert\expect;
 /**
@@ -15,7 +17,7 @@ use function bovigo\assert\expect;
  *
  * @group  predicate
  */
-class CallablePredicateTest extends \PHPUnit_Framework_TestCase
+class CallablePredicateTest extends TestCase
 {
     /**
      * helper method for the test
