@@ -8,6 +8,7 @@ declare(strict_types=1);
  */
 namespace bovigo\assert\predicate;
 use bovigo\assert\AssertionFailure;
+use PHPUnit\Framework\TestCase;
 
 use function bovigo\assert\assert;
 use function bovigo\assert\expect;
@@ -16,7 +17,7 @@ use function bovigo\assert\expect;
  *
  * @group  predicate
  */
-class EqualsTest extends \PHPUnit_Framework_TestCase
+class EqualsTest extends TestCase
 {
     /**
      * @return  array
@@ -104,7 +105,7 @@ additional info"
 --- Expected
 +++ Actual
 @@ @@
--'foo
+-'foo\\n
 -'
 +'bar'
 
