@@ -8,6 +8,7 @@ declare(strict_types=1);
  */
 namespace bovigo\assert\predicate;
 use bovigo\assert\AssertionFailure;
+use PHPUnit\Framework\TestCase;
 
 use function bovigo\assert\assert;
 use function bovigo\assert\expect;
@@ -69,7 +70,7 @@ class IteratorAggregateExample implements \IteratorAggregate
  *
  * @group  predicate
  */
-class IsOfSizeTest extends \PHPUnit_Framework_TestCase
+class IsOfSizeTest extends TestCase
 {
     /**
      * @test
